+++ components/Form/LinkForm.js
@@ -4,10 +4,10 @@ import styled from 'styled-components/native'
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view'
 import { isURL } from 'validator'
 
-import FieldSet from './FieldSet'
-import HeaderRightButton from './HeaderRightButton'
+import FieldSet from '../FieldSet'
+import HeaderRightButton from '../HeaderRightButton'
 
-import { Units } from '../constants/Style'
+import { Units } from '../../constants/Style'
 
 const Container = styled(KeyboardAwareScrollView)`
   flex: 1;
@@ -18,13 +18,17 @@ const Field = styled(FieldSet)`
   margin-top: ${Units.scale[4]};
 `
 
-export default class TextForm extends React.Component {
+export default class LinkForm extends React.Component {
   constructor(props) {
     super(props)
     this.state = {
-      title: props.block.title,
-      description: props.block.description,
-      content: props.block.content,
+      source_url: props.block.source_url,
+    }
+  }
+
+  componentWillReceiveProps(nextProps) {
+    if (nextProps.block) {
+      this.setState({ source_url: nextProps.block.source_url })
     }
   }
 
@@ -61,31 +65,16 @@ export default class TextForm extends React.Component {
   render() {
     return (
       <Container>
-        <FieldSet
-          label="Text"
-          onChange={this.onFieldChange}
-          fields={[
-            {
-              key: 'content',
-              placeholder: 'Text',
-              type: 'textarea',
-              value: this.state.content,
-            },
-          ]}
-        />
         <Field
-          label="Title / Description"
+          isFirst
+          label="Link"
           onChange={this.onFieldChange}
           fields={[
             {
-              key: 'title',
-              placeholder: 'Title',
-              value: this.state.title,
-            },
-            {
-              key: 'description',
-              placeholder: 'Description',
-              value: this.state.description,
+              key: 'source_url',
+              placeholder: 'URL',
+              type: 'url',
+              value: this.state.source_url,
             },
           ]}
         />
@@ -94,25 +83,21 @@ export default class TextForm extends React.Component {
   }
 }
 
-TextForm.propTypes = {
+LinkForm.propTypes = {
   onSubmit: PropTypes.func,
   submitText: PropTypes.string,
   navigation: PropTypes.any,
   navigationOptions: PropTypes.any.isRequired,
   block: PropTypes.shape({
-    title: PropTypes.string,
-    description: PropTypes.string,
-    content: PropTypes.string,
+    source_url: PropTypes.string,
   }),
 }
 
-TextForm.defaultProps = {
+LinkForm.defaultProps = {
   onSubmit: () => null,
   navigation: () => null,
   submitText: 'Done',
   block: {
-    title: '',
-    description: '',
-    content: '',
+    source_url: '',
   },
 }
